main/requirementsVercel.txt
asgiref
attrs
autobahn
Automat
beautifulsoup4
Brotli
certifi
cffi
channels
charset-normalizer
constantly
cryptography

defusedxml
django
django-allauth
django-autoslug
django-ckeditor
django-cors-headers
django-js-asset
djangorestframework
djangorestframework-camel-case
djangorestframework-simplejwt
drf-spectacular
easy-thumbnails
gunicorn
hyperlink
idna
incremental
inflection
jsonschema
jsonschema-specifications
oauthlib
Pillow
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
python3-openid
pytz
PyYAML
referencing
requests
requests-oauthlib
rpds-py
service-identity
six
soupsieve
sqlparse
Twisted
txaio
typing_extensions
uritemplate
urllib3
whitenoise
zope.interface
```
